Make stack hold bytes

+++ ch04/exercises/src/stack/stack_test.go
@@ -3,7 +3,7 @@ package stack
 import "testing"
 
 func TestShouldPopAfterPush (t *testing.T) {
-	stack := Stack{elements: make([] int, 0), count: 0}
+	stack := Stack{elements: make([] byte, 0), count: 0}
 	Push(&stack, 5)
 	t.Log("Pushed 5 to stack")
 	Pop(&stack)
